requirements.txt
streamlit
Pillow
langchain
langchain-community
langchain-groq
